requirements.txt
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
dnspython
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
kafka-python==2.0.2
keyring==10.6.0
keyrings.alt==3.0
mysql-connector-python==8.0.25
pydantic==1.8.2
PyJWT==1.5.3
pymongo==3.11.4
PyMySQL==1.0.2
pyOpenSSL==17.5.0
pyserial==3.4
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
flake8~=3.7.8

